Write one table or multiple tables to a 'GraphPad Prism' '.pzfx' file

Write one table or multiple tables to a 'GraphPad Prism' '.pzfx' file. A table can be a 'matrix', a 'data.frame', or a 'tibble'. All elements of the table should be numeric.

write_pzfx(x, path, row_names = TRUE, x_col = NA)

Arguments

x
Input table or named list of tables that will be 'Data Tables' in the '.pzfx' file

path
Path to the output '.pzfx' file.

row_names
Logical. If row names of the input table should be preserved and become row titles in the output '.pzfx' file. If the length is greater than 1, it must match the length of list of input tables. Default: TRUE.

x_col
1-based column index or name of the column to be used as the 'X' column. If the length is greater than 1, it must match the length of list of input tables. All other columns in the input tables will be treated as "Y" columns in the output '.pzfx' file. Default: NA

Value

write_pzfx returns the input x invisibly.

Examples

pzfx_file <- system.file("extdata/exponential_decay.pzfx", package = "pzfx", mustWork = TRUE)
df <- read_pzfx(pzfx_file, table = 1, strike_action = "exclude")
write_pzfx(df, path = tempfile(fileext = ".pzfx"), row_names = TRUE)