Read one table from a 'GraphPad Prism' '.pzfx' file

Read one table from a 'GraphPad Prism' '.pzfx' file

read_pzfx(path, table = 1, strike_action = "exclude")

Arguments

path
Path to the '.pzfx' file.

table
Table to read. Either a string (the name of a table), or an integer (the position of the table). If neither argument specifies the table, defaults to the first table.

strike_action
One of c("exclude", "keep", "star") or c("e", "k", "s"). Should stricken values in the original .pzfx be excluded, kept or labeled with a trailing "*". If a trailing "*" is added, the column will be of type character.

Value

a data frame

Examples

pzfx_file <- system.file("extdata/exponential_decay.pzfx", package = "pzfx", mustWork = TRUE)
read_pzfx(pzfx_file, table = 1, strike_action="exclude")

#>    Minutes Control_1 Control_2 Control_3 Treated_1 Treated_2 Treated_3
#> 1        1      8887      7366      9612      6532      7905      7907
#> 2        2      8329        NA      8850      5352      5841      6277
#> 3        3      7907      8810      8669      5177      4082      3157
#> 4        4      7413      8481      6489      3608        NA      4226
#> 5        5      7081      7178      5716      2559      3697      2816
#> 6        6      6249      6492        NA      1671      3053      2891
#> 7        8      5442      6172      6409      2264      1658      1879
#> 8       10      4020      3758      4138      1905      1302      1406
#> 9       14      4559      3146      2547      2994      1338       739
#> 10      20      3033      1587      2754      1444        NA       760
#> 11      25      2105      1707      2152       281       484       765
#> 12      30      1005      2156      1185      1103      1517       833
#> 13      50       820      1513      1591      1918      1128      1293